% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/url.R
\name{check_profile}
\alias{check_profile}
\alias{query_params_list}
\alias{is_correct_query}
\alias{profile_from_query}
\title{Check student profile to identify it from URL's parameters}
\usage{
check_profile(profile, url_profile)

query_params_list()

is_correct_query(query)

profile_from_query(query)
}
\arguments{
\item{profile}{Profile data gathered from the LRS}

\item{url_profile}{Profile data gathered from the URL}

\item{query}{The query part of the URL (text after ?, like in
https://mysite.org?params1=1&param2=2). The query is a character string with
each value, named with the parameters names.}
}
\value{
\code{TRUE} if the profile is correct, \code{FALSE} otherwise.

A character vector with the list of parameters

\code{TRUE} if all required parameters are in the query, \code{FALSE} otherwise.

A list with the required parameters
}
\description{
Check if a GitHub user profile exists by checking its homepage.

Get the list of required parameters to identify a student from the URL.
}
\details{
In Moodle, the URL variables must be set as follows:
\itemize{
\item iemail        = Student's email
\item iid           = Student' ID ("Nom d'utilisateur" in French in Moodle)
\item ifirstname    = First name of the student
\item ilastname     = Last name of the student
\item institution  = Institution
\item icourse      = Identifier for the course
\item ictitle      = Title of the course
\item iurl         = URL of the Moodle server
\item iref         = Moodle ID
}
}
\examples{
query_params_list()
is_correct_query(c(iemail = "student@uni.edu", iid = "1234",
  ifirstname = "John", ilastname = "Doe", institution = "University",
  icourse = "123", ictitle = "Course title", iurl = "https://moodle.uni.edu",
  iref = "2FAB3EE8-42B5-4B61-DA23-2167FACD7B569"))
is_correct_query(c(iemail = "student@uni.edu"))
profile_from_query(c(iemail = "student@uni.edu", iid = "1234", other = "a"))
}
