% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lrs.R
\name{get_data}
\alias{get_data}
\alias{get_profile}
\alias{get_all_data}
\alias{get_github_data}
\alias{get_grid_data}
\alias{get_grade_data}
\alias{get_grade_dir}
\title{Get data from exercises or projects for one student}
\usage{
get_data(
  collection,
  query,
  fields = "{}",
  url = getOption("learnitr.lrs_url"),
  count = FALSE
)

get_profile(email, login = NULL)

get_all_data(email, icourse, module = NULL)

get_github_data(email)

get_grid_data(email, icourse, module = NULL)

get_grade_data(email, icourse)

get_grade_dir(root_dir = "/data1/grades", acad_year = NULL)
}
\arguments{
\item{collection}{The collection in the MongoDB database}

\item{query}{The JSON query on the MongoDB database}

\item{fields}{The fields to return}

\item{url}{The url to access the MongoDB database}

\item{count}{Do we count items instead of getting them?}

\item{email}{The email of the student}

\item{login}{The GitHub login of the student as alternate way to get its
profile (when it is not \code{NULL})}

\item{icourse}{The course identifier}

\item{module}{The module, or \code{NULL} for exercises in all modules}

\item{root_dir}{The root directory where the grades are stored.}

\item{acad_year}{The academic year to use, or \code{NULL} to use the latest one.}
}
\value{
A data frame or a list with the data
}
\description{
Get data from exercises or projects for one student
}
