% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github.R
\name{is_github_user}
\alias{is_github_user}
\title{Check a GitHub account}
\usage{
is_github_user(user)
}
\arguments{
\item{user}{The user login to check on GitHub.}
}
\value{
\code{TRUE} if the user login exists on GitHub, \code{FALSE} otherwise.
}
\description{
Check a GitHub account
}
\examples{
is_github_user("phgrosjean")
is_github_user("nonexistinguser")
}
