% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learnitprogress-package.R
\docType{package}
\name{learnitprogress-package}
\alias{learnitprogress}
\alias{learnitprogress-package}
\title{learnitprogress: Report Student Progress in 'learnitr' Courses}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

A Shiny app that connects to you LRS ("Learning Record Store" database, as created and managed by the 'learnitdown' package) and displays a report showing the completion of exercices and projects. This report is intended for students to track their progress in the course.
}
\details{
The learnitprogress package implements a Shiny application that summarizes
the progression of each of your course's students in their LearnIt::R
exercises and projects. It queries the MongoDB database that is collecting
your LearnIt::R activities (H5P, Shiny, learnr, GitHub projects...).

See ...
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/learnitr/learnitprogress}
  \item \url{https://learnitr.github.io/learnitprogress/}
  \item \url{https://learnitr.r-universe.dev/learnitprogress}
  \item Report bugs at \url{https://github.com/learnitr/learnitprogress/issues}
}

}
\author{
\strong{Maintainer}: Philippe Grosjean \email{phgrosjean@sciviews.org} (\href{https://orcid.org/0000-0002-2694-9471}{ORCID})

}
\keyword{internal}
