% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_grade.R
\name{plot_grade}
\alias{plot_grade}
\title{Visualize details about the grade for one student}
\usage{
plot_grade(data, show.name = TRUE, give.note = TRUE)
}
\arguments{
\item{data}{Grading data}

\item{show.name}{Indicate the name of the student in the title}

\item{give.note}{Indicate the final note in the title}
}
\value{
A ggplot object
}
\description{
Visualize details about the grade for one student
}
