% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_progression.R
\name{plot_progression}
\alias{plot_progression}
\title{Progression plot for one student (exercises H5P, Shiny & Learnr)}
\usage{
plot_progression(user, data)
}
\arguments{
\item{user}{The user login}

\item{data}{The data for this user}
}
\value{
A ggplot object
}
\description{
Progression plot for one student (exercises H5P, Shiny & Learnr)
}
