% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_progress.R
\name{report_progress}
\alias{report_progress}
\alias{run_progress_report}
\title{Show the progress report for one student}
\usage{
report_progress(
  email,
  second.course = FALSE,
  course = NULL,
  module = NULL,
  browse = TRUE,
  url = getOption("learnitr.lrs_url"),
  port = 3260
)

run_progress_report(port = 3260)
}
\arguments{
\item{email}{The email address of the student}

\item{second.course}{Should we display data for a secondary course that the
student follows?}

\item{course}{The primary course}

\item{module}{Should we restrict data for one module, (or all with \code{NULL})?}

\item{browse}{Should we browse to the provided URL?}

\item{url}{The url to connect to the MongoDB database}

\item{port}{The port where the Shiny application is serving. If 0, a version
on a server (for instance, on Posit Connect) is used.}
}
\value{
The URL to the progress report
}
\description{
Show the progress report for one student
}
\examples{
\dontrun{
# In a different R session, run the following line:
learnitprogress::run_progress_report()
# Then...
options(learnitr.lrs_url = "mongodb://127.0.0.1/sdd")
learnitprogress::report_progress("student.email@school.edu")
}
}
